feat(inventory): refactor inventory service and add concurrency support feat(engine): update stock handling with new inventory methods feat(storage): implement event logging and event structure

+++ core/stock.go
@@ -3,5 +3,5 @@ package core
 type StockItem struct {
 	ProductID string
 	Quantity  float64
-	Cost      int
+	Cost      int64
 }
